+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - feat: Send SDK integrations
+- fix: Don't track OOMs for unit tests (#1651)
+- fix: Add verification for vendor UUID in OOM logic (#1648)
 
 ## 7.9.0